LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/properties-value-auto-001-expected.txt


FAIL margin-top auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-top auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-top:2s" but got ""
FAIL margin-top auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL margin-top auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-top:2s" but got ""
FAIL margin-right auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-right auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-right:2s" but got ""
FAIL margin-right auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL margin-right auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-right:2s" but got ""
FAIL margin-bottom auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-bottom auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-bottom:2s" but got ""
FAIL margin-bottom auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL margin-bottom auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-bottom:2s" but got ""
FAIL margin-left auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL margin-left auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-left:2s" but got ""
FAIL margin-left auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL margin-left auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "margin-left:2s" but got ""
FAIL height auto(to) / values assert_not_equals: must not be target value after start got disallowed value "18px"
FAIL height auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "height:2s" but got ""
FAIL height auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.328125px"
FAIL height auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "height:2s" but got ""
FAIL width auto(to) / values assert_not_equals: must not be target value after start got disallowed value "100px"
FAIL width auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "width:2s" but got ""
FAIL width auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.328125px"
FAIL width auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "width:2s" but got ""
FAIL clip auto(to) / values assert_not_equals: must not be target value after start got disallowed value "auto"
PASS clip auto(to) / events
PASS clip auto(from) / values
PASS clip auto(from) / events
FAIL marker-offset auto(to) / values assert_not_equals: initial and target values may not match got disallowed value ""
FAIL marker-offset auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "marker-offset:2s" but got ""
FAIL marker-offset auto(from) / values assert_not_equals: initial and target values may not match got disallowed value ""
FAIL marker-offset auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "marker-offset:2s" but got ""
FAIL top auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL top auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "top:2s" but got ""
FAIL top auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL top auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "top:2s" but got ""
FAIL right auto(to) / values assert_not_equals: must not be target value after start got disallowed value "22.46875px"
FAIL right auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "right:2s" but got ""
FAIL right auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL right auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "right:2s" but got ""
FAIL left auto(to) / values assert_not_equals: must not be target value after start got disallowed value "0px"
FAIL left auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "left:2s" but got ""
FAIL left auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL left auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "left:2s" but got ""
FAIL bottom auto(to) / values assert_not_equals: must not be target value after start got disallowed value "82px"
FAIL bottom auto(to) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "bottom:2s" but got ""
FAIL bottom auto(from) / values assert_not_equals: must not be target value after start got disallowed value "13.333333015441895px"
FAIL bottom auto(from) / events assert_equals: Expected TransitionEnd events triggered on .transition expected "bottom:2s" but got ""
PASS z-index auto(to) / values
PASS z-index auto(to) / events
PASS z-index auto(from) / values
PASS z-index auto(from) / events
Text sample
